implement basic music player
tried to tie into relevant methods for starting at the correct time, and
for seeking forwards in the optimal way. this is kinda complex logic so
will probably write out some tests for it. not sure if this approach is
practical, though, and tests won't really help with validating that. can
try to worsen my internet connection

still todo:
test the player
make things look nice
make mixes on soundcloud
maybe add a thing on the page that makes it less boring to look at
write a readme
some things that i forgot and will take forever
SHIP IT