move visualizer into react, strongly type it
my god this is so much nicer. i guess i could have done this from the
start but when i was figuring out if processing was even a good choice
it was a little less clear this would be worth the effort. so nice to
have types in this, and be able to pass it props is even nicer

still todo:
make mixes on soundcloud
write a readme
some other things i probably forgot about
figure out what SHIP IT means
SHIP IT